First version with full Spring configuration,  almost 0 XML is left (only configurable-context and db)
