+++ src/main/java/com/graffitab/server/config/web/WebConfig.java
@@ -1,25 +1,17 @@
 package com.graffitab.server.config.web;
 
 import java.util.List;
-import java.util.Properties;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.ComponentScan.Filter;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.FilterType;
-import org.springframework.context.annotation.ImportResource;
 import org.springframework.core.annotation.Order;
-import org.springframework.core.io.ClassPathResource;
 import org.springframework.http.converter.HttpMessageConverter;
 import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
-import org.springframework.jdbc.datasource.IsolationLevelDataSourceAdapter;
-import org.springframework.orm.hibernate4.HibernateTransactionManager;
-import org.springframework.orm.hibernate4.LocalSessionFactoryBean;
 import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Controller;
-import org.springframework.transaction.TransactionDefinition;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.multipart.commons.CommonsMultipartResolver;
@@ -30,20 +22,19 @@ import org.springframework.web.servlet.mvc.method.annotation.RequestMappingHandl
 
 import com.graffitab.server.config.spring.CustomMappingJackson2HttpMessageConverter;
 import com.graffitab.server.config.spring.JsonDtoArgumentResolver;
-import com.mchange.v2.c3p0.ComboPooledDataSource;
 
 @Configuration
 @EnableWebMvc
 @EnableTransactionManagement
 @ComponentScan(
-  basePackages={"com.graffitab.server"}, useDefaultFilters = true 
+  basePackages={"com.graffitab.server.api"}, useDefaultFilters = false, 
+  includeFilters = { 
+	      @Filter(type = FilterType.ANNOTATION, classes = {Controller.class}),
+	      @Filter(type = FilterType.ANNOTATION, classes = {Component.class})
+  }
 )
-@ImportResource("classpath:spring/spring-context.xml")
 @Order(2)
 public class WebConfig extends WebMvcConfigurationSupport {
-	
-	@Autowired
-	private ComboPooledDataSource targetDataSource;
 	
 	@Bean
 	public RequestMappingHandlerAdapter requestMappingHandlerAdapter() {
@@ -102,41 +93,5 @@ public class WebConfig extends WebMvcConfigurationSupport {
 		argumentResolvers.add(jsonDtoArgumentResolver());
 	}
 	
-	@Bean
-	public HibernateTransactionManager transactionManager() {
-		HibernateTransactionManager transactionManager = new HibernateTransactionManager();
-		transactionManager.setSessionFactory(sessionFactory().getObject());
-		return transactionManager;
-	}
 	
-	
-	@Bean
-	public IsolationLevelDataSourceAdapter dataSource() {
-		IsolationLevelDataSourceAdapter dataSource = new IsolationLevelDataSourceAdapter();
-		dataSource.setIsolationLevel(TransactionDefinition.ISOLATION_READ_COMMITTED);
-		dataSource.setTargetDataSource(targetDataSource);
-		return dataSource;
-	}
-	
-	@Bean
-    public LocalSessionFactoryBean sessionFactory() {
-		
-        LocalSessionFactoryBean sessionFactory = new LocalSessionFactoryBean();
-        sessionFactory.setDataSource(dataSource());
-        sessionFactory.setMappingDirectoryLocations(new ClassPathResource("hibernate-mappings"));
-        sessionFactory.setHibernateProperties(hibernateProperties());
-        return sessionFactory;
-    }
-	
-	Properties hibernateProperties() {
-        return new Properties() {
-			private static final long serialVersionUID = 1L;
-
-			{
-                setProperty("hibernate.dialect", "org.hibernate.dialect.MySQLInnoDBDialect");
-                setProperty("hibernate.show_sql", "false");
-                setProperty("hibernate.hbm2ddl.auto", "validate");
-            }
-        };
-    }
 }
