--- src/main/resources/spring/persistence.xml
@@ -1,42 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xmlns:tx="http://www.springframework.org/schema/tx"
-       xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.1.xsd
-       		http://www.springframework.org/schema/tx http://www.springframework.org/schema/tx/spring-tx-3.1.xsd">
-
-	<tx:annotation-driven/>
-    
-    <!-- Default Hibernate Transaction Manager -->
-    <bean id="transactionManager" class="org.springframework.orm.hibernate4.HibernateTransactionManager">
-        <property name="sessionFactory" ref="sessionFactory"/>
-    </bean>
-
-    <!-- Activates the annotation driven transactions -->
-    <tx:annotation-driven transaction-manager="transactionManager" proxy-target-class="true"/>
-
- 	<bean id="sessionFactory" class="org.springframework.orm.hibernate4.LocalSessionFactoryBean">
-        <property name="dataSource" ref="dataSource"/>
-        <property name="hibernateProperties" ref="hibernateProperties"/>
-        <property name="mappingResources" ref="hibernateMappings"/>
-    </bean>
-
-    <!--  Hibernate Mapping files -->
-    <bean id="hibernateMappings" class="org.springframework.beans.factory.config.ListFactoryBean">
-        <property name="sourceList">
-            <list>
-                <value>hibernate-mappings/User.hbm.xml</value>
-                <value>hibernate-mappings/Cover.hbm.xml</value>
-                <value>hibernate-mappings/Avatar.hbm.xml</value>
-            </list>
-        </property>
-    </bean>
-
-    <bean class="org.springframework.dao.annotation.PersistenceExceptionTranslationPostProcessor"/>
-    
-    <!-- DAOs -->
-    <bean id="userDao" class="com.graffitab.server.persistence.dao.HibernateDaoImpl">
-        <property name="entityClass" value="com.graffitab.server.persistence.model.User"/>
-    </bean>
-
-</beans>
