--- src/main/resources/spring/spring-context.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xmlns:context="http://www.springframework.org/schema/context"
-    xmlns:task="http://www.springframework.org/schema/task"
-    xsi:schemaLocation="http://www.springframework.org/schema/beans
-               			http://www.springframework.org/schema/beans/spring-beans-3.2.xsd
-               			http://www.springframework.org/schema/context
-               			http://www.springframework.org/schema/context/spring-context-3.1.xsd
-            			http://www.springframework.org/schema/task
-            			http://www.springframework.org/schema/task/spring-task-3.0.xsd">
-
-<!--      <context:annotation-config/> -->
-	
-<!--      <context:component-scan base-package="com.graffitab.server" use-default-filters="false"> -->
-<!--          <context:exclude-filter type="annotation" expression="org.springframework.web.bind.annotation.RestController" /> -->
-<!--      </context:component-scan> -->
-     
-
-    <bean id="multipartResolver" class="org.springframework.web.multipart.commons.CommonsMultipartResolver" />
-
-     <!--  Async and scheduling -->
-<!--     <task:annotation-driven /> -->
-<!--     <bean id="jacksonHttpMessageConverter" class="org.springframework.http.converter.json.MappingJackson2HttpMessageConverter"/> -->
-    
-    <import resource="classpath:jdbc.xml"/>
-    <import resource="classpath:configurable-context.xml"/>
-<!--     <import resource="classpath:spring/persistence.xml"/> -->
-<!--     <import resource="classpath:spring/security.xml"/> -->
-
-</beans>
