+++ src/main/webapp/WEB-INF/web.xml
@@ -7,13 +7,6 @@
 
     <display-name>graffitab</display-name>
 
-    <!-- Log4j startup -->
-<!--
-    <listener>
-        <listener-class>com.morenware.dutils.logging.Log4jInitialiserListener</listener-class>
-    </listener>
--->
-
     <servlet>
         <servlet-name>graffitab</servlet-name>
         <servlet-class>org.springframework.web.servlet.DispatcherServlet</servlet-class>
@@ -33,15 +26,38 @@
         <load-on-startup>1</load-on-startup>
     </servlet>
 
-    <!-- ContextLoader declaration -->
+    <!-- Configure ContextLoaderListener to use JavaConfigWebApplicationContext
+         instead of the default XmlWebApplicationContext -->
+    <context-param>
+        <param-name>contextClass</param-name>
+        <param-value>
+            org.springframework.web.context.support.AnnotationConfigWebApplicationContext
+        </param-value>
+    </context-param>
+
+    <!-- Configuration locations must consist of one or more comma- or space-delimited
+         fully-qualified @Configuration classes -->
+    <context-param>
+        <param-name>contextConfigLocation</param-name>
+        <param-value>com.graffitab.server.config.MainConfig</param-value>
+    </context-param>
+
+    <!-- Bootstrap the root application context as usual using ContextLoaderListener -->
     <listener>
         <listener-class>org.springframework.web.context.ContextLoaderListener</listener-class>
     </listener>
 
-    <context-param>
-        <param-name>contextConfigLocation</param-name>
-        <param-value>classpath:spring/spring-context.xml</param-value>
-    </context-param>
+
+
+    <!-- ContextLoader declaration -->
+<!--     <listener> -->
+<!--         <listener-class>org.springframework.web.context.ContextLoaderListener</listener-class> -->
+<!--     </listener> -->
+
+<!--     <context-param> -->
+<!--         <param-name>contextConfigLocation</param-name> -->
+<!--         <param-value>classpath:spring/spring-context.xml</param-value> -->
+<!--     </context-param> -->
 
     <!-- Filter to force encoding to UTF 8 -->
     <filter>
